requirements.txt
#adium-theme-ubuntu==0.3.4
asn1crypto==0.24.0
#asymmetric-jwt-auth==0.4.1
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
certifi==2017.11.5
cffi==1.11.3
chardet==3.0.4
click==6.7
cov-core==1.15.0
coverage==4.4.2
cryptography==2.1.4
decorator==4.2.1
Django==1.11.9
django-allauth==0.34.0
django-cors-headers==2.1.0
django-debug-toolbar==1.9.1
django-filter==1.1.0
django-mssql==1.8
django-oauth-plus==2.2.6
django-oauth2-provider==0.2.6.1
django-rest-auth==0.9.2
django-sqlserver==1.11
django-tastypie==0.14.0
django-tastypie-hmacauth==0.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
enum34==1.1.6
flake8==2.4.0
ghp-import==0.5.5
httplib2==0.10.3
idna==2.6
ipaddress==1.0.19
ipython==5.5.0
ipython-genutils==0.2.0
Jinja2==2.10
livereload==2.5.1
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.3.1
#mkdocs==0.13.2
MySQL-python==1.2.5
oauth2==1.5.211
oauthlib==2.0.6
pathlib2==2.3.0
pbr==3.1.1
pep8==1.5.7
pexpect==4.3.1
pickleshare==0.7.4
pkginfo==1.4.1
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
py==1.5.2
pyasn1==0.4.2
pycparser==2.18
pycryptodome==3.4.8
pyflakes==0.8.1
Pygments==2.2.0
PyJWT==1.5.3
pymssql==2.1.3
PyMySQL==0.8.0
pyOpenSSL==17.5.0
pytest==2.6.4
pytest-cov==1.6
pytest-django==2.8.0
python-dateutil==2.6.1
python-mimeparse==1.6.0
python-openid==2.2.5
python-tds==1.9.0
pytz==2017.3
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
scandir==1.6
shortuuid==0.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
South==1.0.2
sqlparse==0.2.4
stevedore==1.28.0
tornado==4.5.3
traitlets==4.3.2
#twine==1.4.0
#unity-lens-photos==1.0
urllib3==1.22
#uWSGI==2.0.16
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.8.2
wcwidth==0.1.7
